ref docs: links, * ex, build
I hadn't compiled the html link that was broken into the actual man
pages.
I added links to `covidcast_epidata` to `covidcast_meta` and
`covidcast`, and an example of the use of * to covidcast.
When I compiled the roxygen2 docs, an extra \details{ appeared, which I
assume is because of a change someone else made that didn't get compiled in.